+++ src/main/java/org/tonkushin/hw05/dao/author/AuthorDaoJdbc.java
@@ -24,22 +24,22 @@ public class AuthorDaoJdbc implements AuthorDao {
     }
 
     @Override
-    public int count() {
+    public long count() {
         Map<String, Object> params = new HashMap<>();
-        return jdbc.queryForObject("select count(*) from authors", params, Integer.class);
+        return jdbc.queryForObject("select count(*) from authors", params, Long.class);
     }
 
     @Override
-    public int insert(Author item) {
+    public long insert(Author item) {
         KeyHolder keyHolder = new GeneratedKeyHolder();
         MapSqlParameterSource params = new MapSqlParameterSource();
         params.addValue("name", item.getName());
         jdbc.update("insert into authors (`name`) values (:name)", params, keyHolder);
-        return keyHolder.getKey().intValue();
+        return keyHolder.getKey().longValue();
     }
 
     @Override
-    public Author getById(int id) {
+    public Author getById(long id) {
         Map<String, Object> params = Collections.singletonMap("id", id);
         return jdbc.queryForObject("select * from authors where id = :id", params, new AuthorMapper());
     }
@@ -51,7 +51,7 @@ public class AuthorDaoJdbc implements AuthorDao {
     }
 
     @Override
-    public void deleteById(int id) {
+    public void deleteById(long id) {
         Map<String, Object> params = Collections.singletonMap("id", id);
         jdbc.update("delete from authors where id = :id", params);
     }
